requirements.txt
acme==1.1.0
amqp==2.6.1
ansible==5.10.0
ansible-core==2.12.8
appdirs==1.4.3
apturl==0.5.2
asgiref==3.5.2
asn1crypto==0.24.0
attrs==19.3.0
autopep8==1.5.4
Babel==2.6.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
billiard==3.6.4.0
blinker==1.4
Brlapi==0.7.0
cachetools==4.1.1
celery==4.4.2
certbot==0.40.0
certbot-nginx==0.40.0
certifi==2021.5.30
cffi==1.14.6
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==2.0.6
Cheetah3==3.2.4
cheroot==8.6.0
CherryPy==18.6.1
click==8.0.3
colorama==0.4.3
command-not-found==0.3
ConfigArgParse==0.13.0
configobj==5.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
cupshelpers==1.0
Cython==0.29.30
dbus-python==1.2.16
debtcollector==2.0.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==1.0.0
dj-static==0.0.6
Django==3.2.6
django-admin-interface==0.22.1
django-auth-ldap==2.0.0
django-bootstrap5==21.2
django-celery-results==1.2.1
django-colorfield==0.7.2
django-cors-headers==3.5.0
django-extensions==3.1.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-icons==21.1
django-login-required-middleware==0.8.0
django-otp==1.1.3
django-phonenumber-field==5.2.0
django-python3-ldap==0.15.3
django-qr-code==2.3.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docutils==0.16
drf-yasg==1.20.0
duplicity==0.8.12.0
entrypoints==0.3
ethtool==0.14
facebook-sdk==3.1.0
Faker==4.1.0
fasteners==0.14.1
filelock==3.7.1
Flask==2.0.2
flower==0.9.7
future==0.18.2
gevent==21.12.0
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
greenlet==1.1.2
grequests==0.6.0
gunicorn==19.6.0
hickory==1.0.0
html5lib==1.0.1
httplib2==0.18.1
humanize==4.2.3
identify==2.5.1
idna==2.8
importlib-metadata==1.5.0
importlib-resources==5.8.0
inflection==0.5.1
ipaddr==2.2.0
IPy==1.0
isc==2.0
iso8601==0.1.12
itsdangerous==2.0.1
itypes==1.2.0
jaraco.classes==3.2.1
jaraco.collections==3.5.1
jaraco.context==4.1.1
jaraco.functools==3.5.0
jaraco.text==3.8.0
Jinja2==3.0.1
jmespath==0.9.4
josepy==1.2.0
jsonschema==3.2.0
keyring==18.0.1
keyrings.alt==3.4.0
kombu==4.6.11
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
ldap3==2.9.1
libsass==0.21.0
libvirt-python==8.4.0
lockfile==0.12.2
louis==3.12.0
lxml==4.9.0
M2Crypto==0.31.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
mock==3.0.5
monotonic==1.5
more-itertools==4.2.0
msgpack==0.6.2
mysqlclient==2.1.1
ndg-httpsclient==0.5.1
netaddr==0.7.19
netifaces==0.10.4
networkx==2.4
nftables==0.1
nodeenv==1.7.0
ntlm-auth==1.1.0
numpy==1.23.1
oauth==1.0.1
oauthlib==3.1.1
olefile==0.46
openapi-codec==1.3.2
oslo.config==8.0.2
oslo.context==3.0.2
oslo.i18n==4.0.1
oslo.log==4.1.1
oslo.serialization==3.1.1
oslo.utils==4.1.1
packaging==21.0
PAM==0.4.2
pandas==1.4.3
paramiko==2.6.0
parsedatetime==2.4
pbr==5.8.1
pcp==5.0
pexpect==4.6.0
phonenumbers==8.12.33
Pillow==8.3.1
pipenv==2022.9.4
platformdirs==2.5.2
ply==3.11
portend==3.1.0
pre-commit==2.19.0
prometheus-client==0.8.0
protobuf==3.13.0
psutil==5.5.1
psycopg2==2.9.3
psycopg2-binary==2.9.1
pusher==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.6.0
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pyflakes==2.1.1
Pygments==2.3.1
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
PyJWT==2.1.0
pykerberos==1.1.14
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyparted==3.11.2
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-decouple==3.5
python-dotenv==0.19.0
python-ldap==3.2.0
python-magic==0.4.16
python-novnc==1.0.0
python-pam==1.8.4
python-twitter==3.5
python3-openid==3.2.0
pytz==2022.2.1
pywinrm==0.3.0
pyxdg==0.26
PyYAML==5.3.1
qrcode==7.3.1
redis==3.5.3
regex==2022.9.13
reportlab==3.5.34
requests==2.21.0
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
requests-unixsocket==0.2.0
resolvelib==0.5.4
rfc3986==1.3.2
roman==2.0.0
rsa==4.6
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
rwlock==0.0.7
schedule==1.1.0
SecretStorage==2.3.1
segno==1.5.2
selinux==3.0
sepolicy==3.0
setools==4.2.2
simplejson==3.16.0
six==1.16.0
slack-sdk==3.19.1
slip==0.6.5
slip.dbus==0.6.5
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sos==4.3
soupsieve==1.9.5
sqlparse==0.4.2
ssh-import-id==5.10
static3==0.7.0
stevedore==1.32.0
supervisor==4.1.0
system-service==0.3
systemd-python==234
tempora==5.0.1
testresources==2.0.1
text-unidecode==1.3
toml==0.10.1
tornado==6.2
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unipath==1.1
uritemplate==3.0.1
urllib3==1.24.3
usb-creator==0.3.7
vboxapi==1.0
vine==1.3.0
virtualenv==20.15.0
virtualenv-clone==0.5.7
wadllib==1.3.3
webencodings==0.5.1
websockify==0.10.0
Werkzeug==2.0.2
whitenoise==6.2.0
wrapt==1.11.2
xkit==0.0.0
xmltodict==0.12.0
zc.lockfile==2.0
zipp==3.6.0
zope.component==4.3.0
zope.event==4.5.0
zope.hookable==5.0.0
zope.interface==4.7.1asgiref==3.4.1
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.13.0
django-crontab==0.7.1
djangorestframework==3.14.0
gunicorn==20.1.0
numpy==1.23.4
pandas==1.5.1
pretty-html-table==0.9.16
pycodestyle==2.7.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.21.0
pytz==2021.1
regex==2022.9.13
six==1.16.0
slack-sdk==3.19.1
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0
asgiref==3.4.1
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.13.0
django-crontab==0.7.1
djangorestframework==3.14.0
gunicorn==20.1.0
numpy==1.23.4
pandas==1.5.1
pretty-html-table==0.9.16
pycodestyle==2.7.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.21.0
pytz==2021.1
regex==2022.9.13
six==1.16.0
slack-sdk==3.19.1
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0
